cheetahml/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
fastapi==0.103.2
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
gevent==23.9.1
geventhttpclient==2.0.11
google==3.0.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
locust==2.21.0
logger==1.4
MarkupSafe==2.1.5
msgpack==1.0.7
packaging==23.2
pluggy==1.4.0
prometheus-fastapi-instrumentator==6.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
py-grpc-prometheus==0.7.0
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.4.4
python-dateutil==2.8.2
pyzmq==25.1.2
redis==5.0.1
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.23.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
-r shared_utils/requirements.txt